(PC-30897)[PRO] refactor: suggested subcategories in DetailsForm
