+++ pro/src/screens/IndividualOffer/DetailsScreen/SuggestedSubcategories/SuggestedSubcategories.module.scss
@@ -1,0 +1,24 @@
+@use "styles/mixins/_rem.scss" as rem;
+@use "styles/variables/_size.scss" as size;
+
+.suggested-subcategories {
+    margin-bottom: rem.torem(24px);
+}
+
+.description{
+    margin-bottom: rem.torem(16px);
+}
+
+.items {
+  gap: rem.torem(16px);
+  display: flex;
+  flex-wrap: wrap;
+}
+
+.suggested-subcategory {
+  width: 100%;
+
+  @media (min-width: size.$tablet) {
+    width: fit-content;
+  }
+}
